Move fixtures around for simplified usage and prevent local poetry changes
